requirements.txt
numpy
opencv-contrib-python
torch
torchvision
mediapipe
flask
requests
pillow
flask_cors